Updating to 58.0.3005.0

+++ src/net/spdy/spdy_no_op_visitor.cc
@@ -20,7 +20,8 @@ net::SpdyHeadersHandlerInterface* SpdyNoOpVisitor::OnHeaderFrameStart(
   return this;
 }
 
-bool SpdyNoOpVisitor::OnUnknownFrame(SpdyStreamId stream_id, int frame_type) {
+bool SpdyNoOpVisitor::OnUnknownFrame(SpdyStreamId stream_id,
+                                     uint8_t frame_type) {
   return true;
 }
